Raise original exception on login failure
The current implementation retries to login in http mode when https
login fails. If http login fails too, then it will raise the exception
of the retry attempt, masking the original error. This change raises the
https loging exception when this operation fails (unless the http login
succeeds) and instead masks the retry exception, if it happens.